docs - update some documentation
