+++ src/Crawler.ts
@@ -1,8 +1,9 @@
 import { FRAME_TYPES } from './constants';
 
-/** WIP
+/**
  * @description A class to handle traversing an array of selected items and return useful items
- * (parent layer, frame, page, etc). It will also find items based on ID (or timestamp).
+ * (child layers, first in selection, layer position, dimensions and position of layer gaps
+ * and overlapping layer negative space).
  *
  * @class
  * @name Crawler
@@ -28,13 +29,13 @@ export default class Crawler {
     return this.array[0];
   }
 
-  /** WIP
-   * @description Uses `setArray` to ensure the selection array is a javascript array. Also
-   * looks into the selection for any groups and pulls out individual layers.
+  /**
+   * @description Looks into the selection array for any groups and pulls out individual layers,
+   * effectively flattening the selection.
    *
    * @kind function
    * @name all
-   * @returns {Object} All items in the array as a javascript array.
+   * @returns {Object} All items (including children) individual in an updated array.
    */
   all() {
     const initialSelection = this.array;
@@ -55,7 +56,14 @@ export default class Crawler {
         let innerLayers = [];
         layer.children.forEach(child => innerLayers.push(child));
 
-        // WIP
+        /**
+         * @description Iterates through `innerLayers`, adding normal layers to the `flatSelection`
+         * array, while adding any additional children layers into the `innerLayers` array.
+         *
+         * @kind function
+         * @name iterateKnownChildren
+         * @private
+         */
         const iterateKnownChildren = () => {
           // iterate through known child layers in `innerLayers`,
           // adding more children to the array as they are found in descendent layers
@@ -91,10 +99,10 @@ export default class Crawler {
     return flatSelection;
   }
 
-  /** WIP
-   * @description Simulates Sketch’s position() object, but for an entire selection,
-   * and keeps the coordinates relative to the frame, ignoring if some of the items
-   * are grouped inside other layers.
+  /**
+   * @description Creates an object containing `x`, `y` coordinates and `width` and `height`
+   * an entire selection, keeping the coordinates relative to the frame, ignoring if some
+   * items are grouped inside other layers.
    *
    * @kind function
    * @name position
@@ -115,7 +123,6 @@ export default class Crawler {
     // iterate through the selected layers and update the frame inner `x`/`y` values and
     // the outer `x`/`y` values
     this.all().forEach((layer) => {
-      // const layerCoordinates = getPositionOnArtboard(layer);
       const layerX = layer.x;
       const layerY = layer.y;
       const layerW = layer.width;
