+++ src/constants.js
@@ -21,13 +21,12 @@ const PLUGIN_IDENTIFIER = 'com.linkedin.figma.specter-plugin';
  */
 const PLUGIN_NAME = 'Specter';
 
-/** WIP
- * @description The public-facing name for the plugin. This should match the
- * `name` stated in manifset.json.
+/**
+ * @description An object containing the set of colors in-use by the plugin.
  *
  * @kind constant
- * @name PLUGIN_NAME
- * @type {string}
+ * @name COLORS
+ * @type {Object}
  */
 const COLORS = {
   component: '#9966ff',
@@ -37,26 +36,25 @@ const COLORS = {
   style: '#ff6655',
 };
 
-/** WIP
- * @description The public-facing name for the plugin. This should match the
- * `name` stated in manifset.json.
+/**
+ * @description An object containing the current string constants the Figma API returns for
+ * top-level (`main`) layer and `group` layer types.
  *
  * @kind constant
- * @name PLUGIN_NAME
- * @type {string}
+ * @name FRAME_TYPES
+ * @type {Object}
  */
 const FRAME_TYPES = {
   main: 'FRAME',
   group: 'GROUP',
 };
 
-/** WIP
- * @description The public-facing name for the plugin. This should match the
- * `name` stated in manifset.json.
+/**
+ * @description An object containing `height`/`width` settings for the plugin GUI window.
  *
  * @kind constant
- * @name PLUGIN_NAME
- * @type {string}
+ * @name GUI_SETTINGS
+ * @type {Object}
  */
 const GUI_SETTINGS = {
   default: {
@@ -65,6 +63,13 @@ const GUI_SETTINGS = {
   },
 };
 
+/**
+ * @description A unique message that will cause the plugin to shut-down.
+ *
+ * @kind constant
+ * @name CLOSE_PLUGIN_MSG
+ * @type {string}
+ */
 const CLOSE_PLUGIN_MSG = '_CLOSE_PLUGIN_';
 
 export {
